requirements.txt
absl-py==0.7.1
appnope==0.1.0
astor==0.7.1
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
decorator==4.4.0
defusedxml==0.6.0
entrypoints==0.3
Flask==1.0.3
gast==0.2.2
grpcio==1.20.1
gunicorn==19.9.0
h5py==2.9.0
idna==2.8
ipykernel==5.1.0
ipython==7.5.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.13.3
Jinja2==2.10.1
jsonschema==3.0.1
jupyter-client==5.2.4
jupyter-core==4.4.0
jupyterlab==0.35.5
jupyterlab-server==0.2.0
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
Markdown==3.1
MarkupSafe==1.1.1
mistune==0.8.4
mock==3.0.5
nbconvert==5.5.0
nbformat==4.4.0
notebook==5.7.8
numpy==1.16.3
opencv-python==3.4.5.20
pandocfilters==1.4.2
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
postgres==2.2.2
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.7.1
psycopg2-binary==2.8.3
ptyprocess==0.6.0
Pygments==2.4.0
pyrsistent==0.14.11
python-dateutil==2.8.0
pyzmq==18.0.0
requests==2.22.0
Send2Trash==1.5.0
six==1.12.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.2
traitlets==4.3.2
urllib3==1.25.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.4
